+++ addons/interaction_system/InteractionSystem.cs
@@ -6,21 +6,21 @@ namespace InteractionSystem
 	[Tool]
 	public partial class InteractionSystem : EditorPlugin
 	{
-		private const string commonPath = "res://addons/interaction_system/assets/textures/icon.png";
+		private const string commonPath = "res://addons/interaction_system/src/";
 
 		public override void _EnterTree()
 		{
-			var texture = GD.Load<Texture2D>("res://icon.png");
+			var icon = GD.Load<Texture2D>("res://addons/interaction_system/assets/textures/icon.png");
 
-			AddCustomType("Interactable2D", "Area2D", GD.Load<Script>(commonPath + "interactable/Interactable2D.cs"), texture);
-			AddCustomType("Interactable3D", "Area3D", GD.Load<Script>(commonPath + "interactable/Interactable3D.cs"), texture);
-			AddCustomType("InteractableProp3D", "Area3D", GD.Load<Script>(commonPath + "interactable/InteractableProp3D.cs"), texture);
+			AddCustomType("Interactable2D", "Node", GD.Load<Script>(commonPath + "interactable/Interactable2D.cs"), icon);
+			AddCustomType("Interactable3D", "Node", GD.Load<Script>(commonPath + "interactable/Interactable3D.cs"), icon);
+			AddCustomType("InteractableProp3D", "Node", GD.Load<Script>(commonPath + "interactable/InteractableProp3D.cs"), icon);
 
-			AddCustomType("Interactor2D", "Node2D", GD.Load<Script>(commonPath + "interactor/Interactor2D.cs"), texture);
-			AddCustomType("Interactor3D", "Node3D", GD.Load<Script>(commonPath + "interactor/Interactor3D.cs"), texture);
+			AddCustomType("Interactor2D", "Node", GD.Load<Script>(commonPath + "interactor/Interactor2D.cs"), icon);
+			AddCustomType("Interactor3D", "Node", GD.Load<Script>(commonPath + "interactor/Interactor3D.cs"), icon);
 
-			AddCustomType("CharacterInteractor2D", "Node2D", GD.Load<Script>(commonPath + "interactor/character_interactor/CharacterInteractor2D.cs"), texture);
-			AddCustomType("CharacterInteractor3D", "Node3D", GD.Load<Script>(commonPath + "interactor/character_interactor/CharacterInteractor3D.cs"), texture);
+			AddCustomType("CharacterInteractor2D", "Node", GD.Load<Script>(commonPath + "interactor/character_interactor/CharacterInteractor2D.cs"), icon);
+			AddCustomType("CharacterInteractor3D", "Node", GD.Load<Script>(commonPath + "interactor/character_interactor/CharacterInteractor3D.cs"), icon);
 		}
 
 		public override void _ExitTree()