+++ addons/interaction_system/src/interactable/Interactable.cs
@@ -2,12 +2,12 @@ using Godot;
 
 namespace InteractionSystem.Interactable
 {
-	public partial interface IInteractable
+	public partial class Interactable : Node
 	{
-		[Signal] public delegate void InteractedEventHandler(Interactor.IInteractor interactor);
-		[Signal] public delegate void ClosestEventHandler(Interactor.IInteractor interactor);
-		[Signal] public delegate void NotClosestEventHandler(Interactor.IInteractor interactor);
-		[Signal] public delegate void FocusedEventHandler(Interactor.IInteractor interactor);
-		[Signal] public delegate void UnfocusedEventHandler(Interactor.IInteractor interactor);
+		[Signal] public delegate void InteractedEventHandler(Interactor.Interactor interactor);
+		[Signal] public delegate void ClosestEventHandler(Interactor.Interactor interactor);
+		[Signal] public delegate void NotClosestEventHandler(Interactor.Interactor interactor);
+		[Signal] public delegate void FocusedEventHandler(Interactor.Interactor interactor);
+		[Signal] public delegate void UnfocusedEventHandler(Interactor.Interactor interactor);
 	}
 }