+++ addons/interaction_system/src/interactable/Interactable3D.cs
@@ -1,13 +1,46 @@
+using System.Collections.Generic;
 using Godot;
 
 namespace InteractionSystem.Interactable
 {
-	public partial class Interactable3D : Area3D, IInteractable
+	[Tool]
+	public partial class Interactable3D : Interactable
 	{
-		[Signal] public delegate void InteractedEventHandler(Interactor.Interactor3D interactor);
-		[Signal] public delegate void ClosestEventHandler(Interactor.Interactor3D interactor);
-		[Signal] public delegate void NotClosestEventHandler(Interactor.Interactor3D interactor);
-		[Signal] public delegate void FocusedEventHandler(Interactor.Interactor3D interactor);
-		[Signal] public delegate void UnfocusedEventHandler(Interactor.Interactor3D interactor);
+		[Export]
+		public Area3D Area
+		{
+			get => _area;
+			set
+			{
+				if (value != _area)
+				{
+					_area = value;
+					UpdateConfigurationWarnings();
+				}
+			}
+		}
+
+		protected Area3D _area;
+
+		public override void _Ready()
+		{
+			Area.SetMeta("interactable", this);
+		}
+
+		public override string[] _GetConfigurationWarnings()
+		{
+			List<string> warnings = new();
+
+			if (_area is null)
+			{
+				var warning = "This node does not have the ability to be interacted with. " +
+					"Please add an Area3D to this node.";
+				warnings.Add(warning);
+			}
+
+			warnings.AddRange(base._GetConfigurationWarnings() ?? System.Array.Empty<string>());
+
+			return warnings.ToArray();
+		}
 	}
 }